Move Effect Area helper to general item helpers
