+++ src/module/item/spell/data.ts
@@ -1,8 +1,9 @@
 import type { SaveType } from "@actor/types.ts";
 import type { BaseItemSourcePF2e, ItemSystemData, ItemSystemSource, ItemTraits } from "@item/base/data/system.ts";
+import type { EffectAreaShape } from "@item/types.ts";
 import type { OneToTen, ValueAndMax, ZeroToThree } from "@module/data.ts";
 import type { DamageCategoryUnique, DamageKind, DamageType, MaterialDamageEffect } from "@system/damage/index.ts";
-import type { EffectAreaShape, MagicTradition, SpellTrait } from "./types.ts";
+import type { MagicTradition, SpellTrait } from "./types.ts";
 
 type SpellSource = BaseItemSourcePF2e<"spell", SpellSystemSource>;
