+++ src/module/item/spell/values.ts
@@ -1,4 +1,3 @@
-const EFFECT_AREA_SHAPES = ["burst", "cone", "cube", "cylinder", "emanation", "line", "square"] as const;
 const MAGIC_TRADITIONS = new Set(["arcane", "divine", "occult", "primal"] as const);
 
-export { EFFECT_AREA_SHAPES, MAGIC_TRADITIONS };
+export { MAGIC_TRADITIONS };
