+++ src/module/item/values.ts
@@ -28,4 +28,6 @@ const ITEM_TYPES = [
     "weapon",
 ] as const;
 
-export { ITEM_TYPES, PC_ITEM_TYPES };
+const EFFECT_AREA_SHAPES = ["burst", "cone", "cube", "cylinder", "emanation", "line", "square"] as const;
+
+export { EFFECT_AREA_SHAPES, ITEM_TYPES, PC_ITEM_TYPES };
