+++ src/module/scene/measured-template-document.ts
@@ -6,7 +6,7 @@ import type {
 } from "@common/abstract/_types.d.mts";
 import type { DocumentFlags } from "@common/data/_types.d.mts";
 import { ItemPF2e } from "@item";
-import type { EffectAreaShape } from "@item/spell/types.ts";
+import type { EffectAreaShape } from "@item/types.ts";
 import type { MeasuredTemplatePF2e } from "@module/canvas/measured-template.ts";
 import type { ItemOriginFlag } from "@module/chat-message/data.ts";
 import type { ChatMessagePF2e } from "@module/chat-message/document.ts";
